schema explorer cleanup 2
