+++ ingresspipe/utils/__init__.py
@@ -1,1 +1,7 @@
-from .general import update_df, execute_google_api_requests, expand_curie_to_uri, expand_curies_in_schema, find_duplicates, uri2label
+from .curie_utils import expand_curie_to_uri, expand_curies_in_schema, extract_name_from_uri_or_curie, uri2label
+from .df_utils import update_df
+from .general import dict2list, find_duplicates, str2list, unlist, visualize
+from .google_api_utils import execute_google_api_requests
+from .load_utils import export_json, load_default, load_json, load_schemaorg
+from .schema_utils import load_schema_into_networkx
+from .validate_utils import validate_class_schema, validate_property_schema, validate_schema
